обработал ошибки в RequestLoan фиче сделал переход на Success screen

+++ app/src/main/java/com/example/a2022_q2_osovskoy/data/model/network/LoanConditionResponse.kt
@@ -8,7 +8,7 @@ data class LoanConditionResponse(
 	val period: Int,
 
 	@field:SerializedName("maxAmount")
-	val maxAmount: Int,
+	val maxAmount: Long,
 
 	@field:SerializedName("percent")
 	val percent: Double
