+++ app/src/main/java/com/example/a2022_q2_osovskoy/domain/entity/loan/LoanCondition.kt
@@ -1,3 +1,3 @@
 package com.example.a2022_q2_osovskoy.domain.entity.loan
 
-data class LoanCondition(val period: Int, val maxAmount: Int, val percent: Double)
+data class LoanCondition(val period: Int, val maxAmount: Long, val percent: Double)
